FOV restriction settings errors corrected

+++ Assets/Scripts/ExperimentPosition.cs
@@ -2,7 +2,7 @@
 [System.Serializable]
 public class ExperimentPosition{
 	public int id;
-	public int posX;
-    public int posY;
-    public int posZ;
+	public float posX;
+    public float posY;
+    public float posZ;
 }
